+++ packages/cargo/jns42-core/src/schema_transforms/inherit.rs
@@ -42,7 +42,7 @@ macro_rules! generate_mod {
        * ```
        *
        */
-      pub fn transform(arena: &mut SchemaArena, key: u32) {
+      pub fn transform(arena: &mut SchemaArena, key: usize) {
         let item = arena.get_item(key);
 
         let Some(sub_keys) = item.$member.clone() else {
@@ -255,7 +255,7 @@ pub mod reference {
    * ```
    *
    */
-  pub fn transform(arena: &mut SchemaArena, key: u32) {
+  pub fn transform(arena: &mut SchemaArena, key: usize) {
     let item = arena.get_item(key);
 
     let Some(sub_key) = item.reference else {
